src/requirements.txt
numpy
pathlib
datetime
cv2
matplotlib
Pillow
scipy
scikit-image